V0.1.9a (Branch convertMatrixes) 2022-09-23
- theMatrix and stepsDetail variables moved to globalVar.js file
- Solving Techniques Functions moved to solvingTechniques.js file
- discarding functions were moved to a new file (discradingTechniques.js)
- With the changes made since v0.1.1 (when the javascript was contained in just one file) main.js file was reduced from more than 900 lines to 146 lines, and disagragated in several lines for organization and code cleanliness.

---
