+++ js/notesZero.js
@@ -1,5 +1,9 @@
 'use strict';
-import * as recurrent from "./theRecurrentFunctions.js"
+import * as recurrent from "./theRecurrentFunctions.js";
+
+////////////////////////////////////////////////////////////////////////////////
+//                       NOTES to ZERO FUNCTIONS                             //
+//////////////////////////////////////////////////////////////////////////////
 
 //Here, it is mark as zero, each cell in the same row, which contains the currentcellvalue as option yet
 const noteZeroRow = (row, currentcellvalue, theMatrixStep) => {
